agentic_travel_planner_2/requirements.txt
google-generativeai
redis
python-dotenv